Move DownloadProgress to payload live_image
It is only used in the live_image task so move it next to the task.
