+++ pyanaconda/modules/payload/payloads/live_image/initialization.py
@@ -27,7 +27,6 @@ from pyanaconda.modules.common.errors.payload import SourceSetupError
 from pyanaconda.modules.common.task import Task
 from pyanaconda.modules.payload.live.utils import get_local_image_path_from_url, \
     get_proxies_from_option, url_target_is_tarfile
-from pyanaconda.modules.payload.live.initialization import DownloadProgress
 from pyanaconda.payload.utils import mount, unmount
 
 from pyanaconda.anaconda_loggers import get_module_logger
@@ -297,3 +296,43 @@ class TeardownInstallationSourceImageTask(Task):
         if not get_local_image_path_from_url(self._url):
             if os.path.exists(self._image_path):
                 os.unlink(self._image_path)
+
+
+class DownloadProgress(object):
+    """Provide methods for download progress reporting."""
+
+    def __init__(self, url, size, report_callback):
+        """Create a progress object for given task.
+
+        :param url: url of the download
+        :type url: str
+        :param size: length of the file
+        :type size: int
+        :param report_callback: callback with progress message argument
+        :type task: callable taking str argument
+        """
+        self.report = report_callback
+        self.url = url
+        self.size = size
+        self._pct = -1
+
+    def update(self, bytes_read):
+        """Download update.
+
+        :param bytes_read: Bytes read so far
+        :type bytes_read:  int
+        """
+        if not bytes_read:
+            return
+        pct = min(100, int(100 * bytes_read / self.size))
+
+        if pct == self._pct:
+            return
+        self._pct = pct
+        self.report("Downloading image %(url)s (%(pct)d%%)" %
+                    {"url": self.url, "pct": pct})
+
+    def end(self):
+        """Download complete."""
+        self.report("Downloading image %(url)s (%(pct)d%%)" %
+                    {"url": self.url, "pct": 100})
